test(server): relocate runs API test to src/__tests__ for discovery
